backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.1.2
python-decouple==3.8
cryptography==41.0.7

# AI/ML
openai==1.3.7
anthropic==0.7.7
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Vector Database
chromadb==0.4.18
sentence-transformers==2.5.1
huggingface-hub==0.20.3

# Communication
requests==2.31.0
aiohttp==3.8.5
python-telegram-bot==20.7
line-bot-sdk==3.5.0
python-socketio==5.10.0

# File Handling
python-magic==0.4.27
Pillow==10.1.0
aiofiles==23.2.1

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
